pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-complaints-bot"
version = "0.1.0"
description = "RAG Complaint-Analysis Chatbot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2",
  "numpy>=1.26",
  "pyarrow>=16.0",
  "scikit-learn>=1.4",
  "sentence-transformers>=3.0",
  "openai>=1.40",
  "pyyaml>=6.0",
  "tqdm>=4.66",
  "streamlit>=1.36",
  "pydantic>=2.8"
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100